新功能:搜索框主题提示

+++ src/views/layout/components/header/header-search/history.vue
@@ -36,6 +36,7 @@ const EMITS_ITEM_CLICK = 'itemClick'
 
 <script setup>
 import { useStore } from 'vuex'
+import { confirm } from '@/libs'
 
 const emits = defineEmits([EMITS_ITEM_CLICK])
 
@@ -45,7 +46,12 @@ const store = useStore()
  * 删除所有记录
  */
 const onDeleteAllClick = () => {
-    store.commit('search/deleteAllHistory')
+    confirm('要删除所有历史记录吗？').then(() => {
+        store.commit('search/deleteAllHistory')
+    })
+    .catch(() => {
+        console.log('取消删除');
+    })
 }
 
 /**
